images for changing the homepage appearance
Trying to fix my landing page. I want it to be a bunch of blog posts - not a site with no links whatsoever to the blog posts? Who decided this was a good idea?

Themes > appearance
I'm still on the "Blank Canvas" and I want it that way. Click the "customize" button.

Oh lord. So many options. I didn't realize I was launching a spaceship today.

Found it!

Themes > appearance > Homepage Settings

From here, change the settings from "Your homepage displays A static page" to "You homepage displays Your latest posts."

Wow. What a lot of explorations and clicking into the settings to make such a simple change. You see, this is the sort of thing that people go through and once they learn it they kind of forget it (the tacit knowledge phenomenon.) I can already tell I'm going to make a brief instructional video about how to quickly set up a Wordpress blog to be a design journal. Because this was frustrating and we can't have that! Get frustrated about your projet maybe- but not about the technicalities of keeping a journal alongside of your deisgn work!